requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.1
nest_asyncio==1.5.8
